requirements.txt
python-dotenv==1.0.1
PyJWT>=2.8,<3
